Ref T11611#OCID 244## Mutex: protect access to `on_start` and `on_finish`
Before this commit, there is no kind of mutual exclusion in between calling
`p.condition` and `manager.on_start`, so two processes can "request" more
resources than the ones really available on the system because of this. This
situation leaded to some "tricks" such as performing some actions at the
condition call, when they should have been done at the `on_start` call.

Right now, everything should be working in a known and expected way for all
the existing plugins and the new ones