Ref T11807#OCID 321## Petition: use own equality and comparisson operators
We have detected some errors at Orcha's petition handler whenever an
`EmptyPetition` was enqueued:

```
 Traceback (most recent call last):
   File "/usr/lib/python3/dist-packages/orcha/lib/manager.py", line 287, in shutdown
     self.manager.shutdown()
   File "/usr/lib/python3/dist-packages/orcha/lib/processor.py", line 314, in _process
     self._internalq.put(p)
   File "/usr/lib/python3.9/queue.py", line 150, in put
     self._put(item)
   File "/usr/lib/python3.9/queue.py", line 236, in _put
     heappush(self.queue, item)
TypeError: '<' not supported between instances of 'EmptyPetition' and 'VagrantPetition'
```

That error was causing some troubles when we request the orchestrator to finish

In addition, we implemented new features for handling SystemD watchdog if we
are running as a service: every 5 seconds, a `WATCHDOG` is sent to notify about
our behavior. If an error occurs in a critical section thread, we notify about
such error and request a service restart, for safety reasons.