Ref T11845#OCID 471## Petition: termination is handled by developers
This is a big update in Orcha which basically breaks backwards compatibility
with older releases. The orchestrator now has a utility method for deciding
how to finish a running petition. Until now, this has been done by issuing a
`SIGINT` but now it is possible to, for example, run a function: all you need
to do is to define `terminate` function.

In addition, the processor has been adjusted for using this new method when
finishing processes.