Ref T12320#OCID 649## Watchdog: make processor independent from petitions
There were some `if/elif/else` branches for checking if the received petition
is of one known kind or another. This is not idiomatic and leads to both
confussion and less escalability as new "internal" petitions will require an
update of the processor.

By following SOLID principles, the logic for handling `WatchdogPetition`
requests is now at `WatchdogManager` class, so the processor behaves exactly
the same way for every petition.

There is a special case though, which is `EmptyPetition`. As such petition is
internally used for shutting down the processor, it will be the only scenario
in which the processor needs to analyze the received petition and behave
different from the other ones. There is no need to handle this from manager
as it has nothing to do with it, reason why this behavior has been kept as is.