Ref T12355#OCID 688## Processor: fix a race condition when calling `condition`
The previous commit moved some of the logic to the manager but introduced a
race condition in between `condition` and `on_start`, as they were using
different locks. This could lead to potential issues such as reserving
resources when they are indeed not free. For example:

  1. `p1.condition(p1)` evaluates to `True` and will use `4/4 GiB`.
  2. `on_start(p1)` is called and should work fine but...
  3. `p2.condition(p2)` evaluates to `True` and will use `4/4 GiB` BUT
     in reality it will be `8/4 GiB`.
  4. `on_start(p2)` is called and will work and will increase used RAM to
     `4 GiB`.
  5. `on_start(p1)` will increase the RAM to `8 GiB` when there is only `4 GiB`
     available.

It could happen also that the mechanism for adjusting the reserved memory is
not protected, leading to two calls setting each one its RAM to `4 GiB` in
total and then, when calling `on_finish` respectively, the resulting value
is `-4 GiB`.